compiler can distinguish between number types, but cannot document vals yet

+++ cmake-build-debug/out.c
@@ -2,8 +2,8 @@
 int main() {
 
 
-float x = 5;
-int pi = 3.14;
+int x = 5;
+float pi = 3.14;
 
 return 0;
 }
